requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.6
Werkzeug==2.3.6
